make TestPasswordTerminalState cross-platform
