+++ pty/pty.go
@@ -29,6 +29,15 @@ type PTY interface {
 	Resize(height uint16, width uint16) error
 }
 
+// PTYWithFlags represents a PTY whose flags can be inspected, in particular
+// to determine whether local echo is enabled.
+type PTYWithFlags interface {
+	PTY
+
+	// EchoEnabled determines whether local echo is currently enabled for this terminal.
+	EchoEnabled() (bool, error)
+}
+
 // New constructs a new Pty.
 func New() (PTY, error) {
 	return newPty()
